Fix project layout and add dependencies
